hospital-mcp/pyproject.toml
[project]
name = "hospital-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.25.0",
    "motor>=3.7.1",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
]
